spinlocks: Replace test-and-set locks by ticket locks
Test-and-set locks don't provide fairness and are non-deterministic
(w.r.t. the prediction of the future holder of a lock).
This can be a performance problem in case of lock contention.

Let's change the implementation to use ticket locks, which guarantee
a fair locking order (FIFO ordering).

Ticket locks have two counters 'owner' and 'next'.
The counter 'owner' hold the ticket number of the current lock owner.
The counter 'next' holds the latest free ticket number.

The lock is free, if both counters are equal.
To acquire the lock, one atomically increments the 'next' counter
to obtain a new ticket.
Afterwards the counter 'owner' is polled unitl it becomes equal to
the new ticket.
To release a lock, one atmically increments the 'owner' counter.

The implementation uses a 32-bit wide struct, which consists of
two 16-bit counters (owner and next).

Signed-off-by: Christoph Muellner <cmuellner@linux.com>