lib: sbi: Issue a PAUSE hint between spins in spin_lock()
RISC-V's Zihintpause extension includes a PAUSE instruction that
is designed to reduce engergy consumption while executing spin-wait
code sequences.  It is encoded as a HINT, which means that it
behaves like a NOP on systems that don't implement Zihintpause.

Binutils supports Zihintpause since 2.36, so this patch uses
the ".insn" directive to keep the code compatible with older
toolchains.

Signed-off-by: Christoph Müllner <christoph.muellner@vrull.eu>